Routing

+++ app/Http/Controllers/V1/ClientSide/Post/PostController.php
@@ -3,13 +3,13 @@
 namespace App\Http\Controllers\V1\ClientSide\Post;
 
 use App\Models\Post;
-use App\Services\V1\ClinetSideService\ClientSideService;
 use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\ClientSide\Post\PostRequest;
+use App\Http\Requests\ClientSide\Post\UpvoteRequest;
 use App\Http\Resources\V1\ClientSide\Post\PostResource;
-use App\Http\Resources\V1\ClientSide\Post\PostCollection;
+use App\Services\V1\ClinetSideService\ClientSideService;
 
 class PostController extends Controller
 {
@@ -22,7 +22,7 @@ class PostController extends Controller
      */
     public function index()
     {
-        return new PostCollection(Post::all());
+        return PostResource::collection(Post::all());
     }
 
     /**
@@ -30,8 +30,9 @@ class PostController extends Controller
      */
     public function store(PostRequest $request)
     {
+        $this->authorize('store');
         $this->clientSideService->storePost($request);
-        return response()->json(['message' => 'Created success', Response::HTTP_CREATED]);
+        return response()->json(['message' => 'Created success'], Response::HTTP_CREATED);
     }
 
     /**
@@ -47,8 +48,9 @@ class PostController extends Controller
      */
     public function update(PostRequest $request, Post $post)
     {
+        $this->authorize('update', $post);
         $this->clientSideService->updatePost($request, $post);
-        return response()->json(['message' => 'Update success', Response::HTTP_CREATED]);
+        return response()->json(['message' => 'Update success'], Response::HTTP_CREATED);
     }
 
     /**
@@ -56,7 +58,26 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('destroy', $post);
+
+        $post->comments()->delete();
+
+        $post->upvotes()->delete();
+
         $post->delete();
+
+        return response()->noContent();
+    }
+
+    public function upvote(UpvoteRequest $request, Post $post)
+    {
+        $this->authorize('upvote', $post);
+
+        $upvote = $request->validated();
+
+        $upvote['user_id'] = $request->user()->id;
+
+        $post->upvotes()->create($upvote);
 
         return response()->noContent();
     }
